% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{counties}
\alias{counties}
\title{US Census Bureau Counties Data}
\format{
\subsection{\code{counties}}{

A data frame with 3,222 rows and 10 columns:
\describe{
\item{geo_id}{The full Census Bureau GEOID of the geography.}
\item{geo_entity_text}{The plain text name of the geographic entity.}
\item{pop_acs23_5yr}{The estimated total population of the geography from the B01003 table of the 2023 5-year American Community Survey (ACS).}
\item{sum_lev_code}{The three-digit summary level code to identify the type of Census geography.}
\item{geo_entity_type_text}{The plain text name of the geography level identified in the \code{sum_lev_code} column.}
\item{state_fips}{The first two digits of the full county \code{fips_code} that identify the state that the county is in.}
\item{fips_code}{The InterNational Committee for Information Technology Standards (INCITS) formerly known as Federal Information Processing Standards (FIPS) code to ensure uniform identification of geographic entities.}
\item{fips_class_code}{Identifies what type of county or county equivalent the geography is.}
\item{func_stat_code}{Identifies the functional status of the geography specified by the \code{fips_code}.}
\item{ns_code}{The National Standard (NS) code used by the US Geological Survey (USGS) for the geography.}
}
}
}
\source{
\url{https://data.census.gov/}
}
\usage{
counties
}
\description{
A reference file of the names, codes, and populations (via the 2025 5-year ACS)
of US Census Bureau counties and county equivalents.
}
\keyword{datasets}
