% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{csas}
\alias{csas}
\title{US Census Bureau CSA Data}
\format{
\subsection{\code{csas}}{

A data frame with 184 rows and 7 columns:
\describe{
\item{geo_id}{The full Census Bureau GEOID of the geography.}
\item{geo_entity_text}{The plain text name of the geographic entity.}
\item{pop_acs23_5yr}{The estimated total population of the geography from the B01003 table of the 2023 5-year American Community Survey (ACS).}
\item{sum_lev_code}{The three-digit summary level code to identify the type of Census geography.}
\item{geo_entity_type_text}{The plain text name of the geography level identified in the \code{sum_lev_code} column.}
\item{geo_code}{The characters after the "US" in the \code{geo_id} that provide a unique identifier within the specified summary level.}
\item{usps_state_abb}{The two digit state abbreviation used by the USPS to identify the state. This is a list-column with a vector of multiple state abbreviations if the CBSA spans multiple states.}
}
}
}
\source{
\url{https://data.census.gov/}
}
\usage{
csas
}
\description{
A reference file of the names, codes, and populations (via the 2025 5-year ACS)
of US Census Bureau combined statistical areas (CSAs).
}
\keyword{datasets}
