% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regions_and_divisions}
\alias{regions_and_divisions}
\title{US Census Bureau Regions and Divisions Data}
\format{
\subsection{\code{regions_and_divisions}}{

A data frame with 13 rows and 6 columns:
\describe{
\item{geo_id}{The full Census Bureau GEOID of the geography}
\item{geo_entity_text}{The plain text name of the geographic entity.}
\item{pop_acs23_5yr}{The estimated total population of the geography from the B01003 table of the 2023 5-year American Community Survey (ACS).}
\item{sum_lev_code}{The three-digit summary level code to identify the type of Census geography.}
\item{geo_entity_type_text}{The plain text name of the geography level identified in the \code{sum_lev_code} column.}
\item{geo_code}{The characters after the "US" in the \code{geo_id} that provide a unique identifier within the specified summary level.}
}
}
}
\source{
\url{https://data.census.gov/}
}
\usage{
regions_and_divisions
}
\description{
A reference file of the names, codes, and populations (via the 2025 5-year ACS)
of US Census Bureau regions and divisions.
}
\keyword{datasets}
