% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{states}
\alias{states}
\title{US Census Bureau States Data}
\format{
\subsection{\code{states}}{

A data frame with 52 rows and 9 columns:
\describe{
\item{geo_id}{The full Census Bureau GEOID of the geography.}
\item{geo_entity_text}{The plain text name of the geographic entity.}
\item{pop_acs23_5yr}{The estimated total population of the geography from the B01003 table of the 2023 5-year American Community Survey (ACS).}
\item{sum_lev_code}{The three-digit summary level code to identify the type of Census geography.}
\item{geo_entity_type_text}{The plain text name of the geography level identified in the \code{sum_lev_code} column.}
\item{usps_state_abb}{The two digit state abbreviation used by the USPS to identify the state.}
\item{region_name}{The plain text name of the Census Bureau region that the state is in. More information on regions can be found in the \code{regions_and_divisions} data set.}
\item{division_name}{The plain text name of the Census Bureau division that the state is in. More information on divisions can be found in the \code{regions_and_divisions} data set.}
\item{fips_code}{The InterNational Committee for Information Technology Standards (INCITS) formerly known as Federal Information Processing Standards (FIPS) code to ensure uniform identification of geographic entities.}
}
}
}
\source{
\url{https://data.census.gov/}
}
\usage{
states
}
\description{
A reference file of the names, codes, and populations (via the 2025 5-year ACS)
of US Census Bureau States.
}
\keyword{datasets}
