% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_change.R
\name{percent_change}
\alias{percent_change}
\title{Calculate Percent Change Between Two Values}
\usage{
percent_change(start_value, end_value)
}
\arguments{
\item{start_value}{A numeric scalar or vector of starting values.}

\item{end_value}{A numeric scalar or vector of ending values.}
}
\value{
A numeric vector of percent changes.
}
\description{
Computes the percent change from a starting value to an ending value.
}
\details{
This function is vectorized and designed for use in tidyverse workflows
such as \code{dplyr::mutate()}.
}
\examples{
percent_change(100, 120)      # 0.20 (20\% increase)
percent_change(c(100, NA), c(110, 120))  # returns c(0.10, NA) with warning

}
